Update README.md
Becky made extensive edits to README.md during the last weeks of May 2022. Besides minor edits, the main changes were to explain that the new GitHub has both PyrEval and PyrEval+CR; to add the two new images of the basic PyrEval and PyrEval+CR pipelines; to reorganize the list of references to all have the same format, and to be in alphabetical order by last name; to reorganize the list of directories to be in the order in which they appear in the folder; to complete and reorganize the list of scripts; to number the launcher steps explicitly in the description of the launcher.